Fix webviews type errors and enforce svelte-check in CI
