+++ .github/workflows/build.yml
@@ -19,6 +19,7 @@ jobs:
       - run: npm ci
       - run: npm run build
       - run: npm run lint
+      - run: npm run check-webviews
       - run: npm run test
       - run: npm run create-package
       - name: Upload coverage to Coveralls
